extract withAxiom to index.ts
