+++ src/index.ts
@@ -1,2 +1,34 @@
+import { NextConfig } from 'next';
+import { EndpointType, getIngestURL } from './config';
 export { reportWebVitals } from './webVitals';
 export { log } from './logger';
+
+export function withAxiom(nextConfig: NextConfig): NextConfig {
+  return {
+    ...nextConfig,
+    rewrites: async () => {
+      const rewrites = await nextConfig.rewrites?.();
+
+      const ingestEndpoint = getIngestURL(EndpointType.webVitals);
+      if (!ingestEndpoint) {
+        return rewrites || []; // nothing to do
+      }
+
+      const axiomRewrites = [
+        {
+          source: '/axiom/web-vitals',
+          destination: ingestEndpoint,
+        },
+      ];
+
+      if (!rewrites) {
+        return axiomRewrites;
+      } else if (Array.isArray(rewrites)) {
+        return rewrites.concat(axiomRewrites);
+      } else {
+        rewrites.afterFiles = rewrites.afterFiles.concat(axiomRewrites);
+        return rewrites;
+      }
+    },
+  };
+}
